requirements.txt
numpy>=1.26
scipy>=1.12
librosa>=0.10.2
soundfile>=0.12.1
pydub>=0.25.1
PyYAML>=6.0.1
opencv-python>=4.10.0
faster-whisper>=1.1.1
tqdm>=4.66.5
ffmpeg-python>=0.2.0
imageio-ffmpeg>=0.6.0
yt-dlp>=2025.3.31
